move utils/serialization.py > serialization.py
