+++ docs/examples/tests/test_request_data.py
@@ -1,7 +1,7 @@
 from examples.request_data.msgpack_request import app as msgpack_app
 
+from starlite.serialization import encode_msgpack
 from starlite.testing import TestClient
-from starlite.utils.serialization import encode_msgpack
 
 
 def test_msgpack_app() -> None:
