+++ docs/lib/reference/serialization.rst
