+++ docs/lib/reference/utils/index.rst
@@ -7,5 +7,4 @@ starlite.utils
     pagination
     predicates
     scope
-    serialization
     sync
