+++ starlite/openapi/controller.py
@@ -9,8 +9,8 @@ from starlite.enums import MediaType, OpenAPIMediaType
 from starlite.exceptions import ImproperlyConfiguredException
 from starlite.handlers import get
 from starlite.response import Response
+from starlite.serialization import encode_json
 from starlite.status_codes import HTTP_404_NOT_FOUND
-from starlite.utils.serialization import encode_json
 
 if TYPE_CHECKING:
     from pydantic_openapi_schema.v3_1_0.open_api import OpenAPI
