+++ starlite/testing/request_factory.py
@@ -12,9 +12,9 @@ from starlite.app import Starlite
 from starlite.connection import Request
 from starlite.enums import HttpMethod, ParamType, RequestEncodingType, ScopeType
 from starlite.handlers.http_handlers import get
+from starlite.serialization import decode_json, encode_json
 from starlite.types import HTTPScope, RouteHandlerType
 from starlite.types.asgi_types import ASGIVersion
-from starlite.utils.serialization import decode_json, encode_json
 
 if TYPE_CHECKING:
     from starlite.datastructures.cookie import Cookie
