+++ tests/openapi/test_integration.py
@@ -3,9 +3,9 @@ from pydantic_openapi_schema.utils import construct_open_api_with_schema_class
 
 from starlite.app import DEFAULT_OPENAPI_CONFIG
 from starlite.enums import OpenAPIMediaType
+from starlite.serialization import decode_json
 from starlite.status_codes import HTTP_200_OK, HTTP_404_NOT_FOUND
 from starlite.testing import create_test_client
-from starlite.utils.serialization import decode_json
 from tests.openapi.utils import PersonController, PetController
