+++ tests/plugins/piccolo_orm/test_piccolo_orm_plugin_integration.py
@@ -3,9 +3,9 @@ from typing import Callable
 from piccolo.testing.model_builder import ModelBuilder
 
 from starlite.plugins.piccolo_orm import PiccoloORMPlugin
+from starlite.serialization import encode_json
 from starlite.status_codes import HTTP_200_OK, HTTP_201_CREATED
 from starlite.testing import create_test_client
-from starlite.utils.serialization import encode_json
 
 from .endpoints import create_concert, retrieve_studio, retrieve_venues, studio, venues
 from .tables import Band, Concert, Manager, RecordingStudio, Venue
